# app/internal/handler/http/oapi/api.gen.go
// Package oapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package oapi

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for GoogleLoginResult.
const (
	GoogleLoginResultAlreadyLinked    GoogleLoginResult = "already_linked"
	GoogleLoginResultInternalError    GoogleLoginResult = "internal_error"
	GoogleLoginResultInvalidToken     GoogleLoginResult = "invalid_token"
	GoogleLoginResultLoginKeyNotFound GoogleLoginResult = "login_key_not_found"
	GoogleLoginResultNotEnabled       GoogleLoginResult = "not_enabled"
	GoogleLoginResultSuccess          GoogleLoginResult = "success"
	GoogleLoginResultUserNotFound     GoogleLoginResult = "user_not_found"
)

// Defines values for Versions.
const (
	VersionsV10 Versions = "v1.0"
)

// AccessTokenResponse defines model for AccessTokenResponse.
type AccessTokenResponse struct {
	// AccessToken the access token
	AccessToken string `json:"access_token"`
}

// GoogleFirstLoginRequest defines model for GoogleFirstLoginRequest.
type GoogleFirstLoginRequest struct {
	// LoginKey the login key
	LoginKey string `json:"login_key"`
}

// GoogleLoginRequest defines model for GoogleLoginRequest.
type GoogleLoginRequest struct {
	// CurrentUrl the url of the page currently being viewed
	CurrentUrl string `json:"current_url"`
}

// GoogleLoginResponse defines model for GoogleLoginResponse.
type GoogleLoginResponse struct {
	// RedirectUrl the Google's login page URL
	RedirectUrl string `json:"redirect_url"`
}

// GoogleLoginResult defines model for GoogleLoginResult.
type GoogleLoginResult string

// Versions defines model for Versions.
type Versions string

// LogoutParams defines parameters for Logout.
type LogoutParams struct {
	XCSRFToken   *string `json:"X-CSRF-Token,omitempty"`
	RefreshToken string  `form:"refresh_token" json:"refresh_token"`
}

// RefreshAccessTokenParams defines parameters for RefreshAccessToken.
type RefreshAccessTokenParams struct {
	XCSRFToken   *string `json:"X-CSRF-Token,omitempty"`
	RefreshToken string  `form:"refresh_token" json:"refresh_token"`
}

// LinkWithGoogleJSONRequestBody defines body for LinkWithGoogle for application/json ContentType.
type LinkWithGoogleJSONRequestBody = GoogleFirstLoginRequest

// LoginWithGoogleJSONRequestBody defines body for LoginWithGoogle for application/json ContentType.
type LoginWithGoogleJSONRequestBody = GoogleLoginRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /auth/logout)
	Logout(w http.ResponseWriter, r *http.Request, params LogoutParams)

	// (GET /auth/oauth/google/callback)
	CallbackFromGoogle(w http.ResponseWriter, r *http.Request)

	// (POST /auth/oauth/google/link)
	LinkWithGoogle(w http.ResponseWriter, r *http.Request)

	// (POST /auth/oauth/google/login)
	LoginWithGoogle(w http.ResponseWriter, r *http.Request)

	// (POST /auth/refresh)
	RefreshAccessToken(w http.ResponseWriter, r *http.Request, params RefreshAccessTokenParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (POST /auth/logout)
func (_ Unimplemented) Logout(w http.ResponseWriter, r *http.Request, params LogoutParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /auth/oauth/google/callback)
func (_ Unimplemented) CallbackFromGoogle(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /auth/oauth/google/link)
func (_ Unimplemented) LinkWithGoogle(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /auth/oauth/google/login)
func (_ Unimplemented) LoginWithGoogle(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /auth/refresh)
func (_ Unimplemented) RefreshAccessToken(w http.ResponseWriter, r *http.Request, params RefreshAccessTokenParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Logout operation middleware
func (siw *ServerInterfaceWrapper) Logout(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params LogoutParams

	headers := r.Header

	// ------------- Optional header parameter "X-CSRF-Token" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CSRF-Token")]; found {
		var XCSRFToken string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-CSRF-Token", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CSRF-Token", valueList[0], &XCSRFToken, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-CSRF-Token", Err: err})
			return
		}

		params.XCSRFToken = &XCSRFToken

	}

	{
		var cookie *http.Cookie

		if cookie, err = r.Cookie("refresh_token"); err == nil {
			var value string
			err = runtime.BindStyledParameterWithOptions("simple", "refresh_token", cookie.Value, &value, runtime.BindStyledParameterOptions{Explode: false, Required: true})
			if err != nil {
				siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "refresh_token", Err: err})
				return
			}
			params.RefreshToken = value

		} else {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "refresh_token"})
			return
		}
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Logout(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CallbackFromGoogle operation middleware
func (siw *ServerInterfaceWrapper) CallbackFromGoogle(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CallbackFromGoogle(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// LinkWithGoogle operation middleware
func (siw *ServerInterfaceWrapper) LinkWithGoogle(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LinkWithGoogle(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// LoginWithGoogle operation middleware
func (siw *ServerInterfaceWrapper) LoginWithGoogle(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.LoginWithGoogle(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RefreshAccessToken operation middleware
func (siw *ServerInterfaceWrapper) RefreshAccessToken(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params RefreshAccessTokenParams

	headers := r.Header

	// ------------- Optional header parameter "X-CSRF-Token" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-CSRF-Token")]; found {
		var XCSRFToken string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "X-CSRF-Token", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-CSRF-Token", valueList[0], &XCSRFToken, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: false})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "X-CSRF-Token", Err: err})
			return
		}

		params.XCSRFToken = &XCSRFToken

	}

	{
		var cookie *http.Cookie

		if cookie, err = r.Cookie("refresh_token"); err == nil {
			var value string
			err = runtime.BindStyledParameterWithOptions("simple", "refresh_token", cookie.Value, &value, runtime.BindStyledParameterOptions{Explode: false, Required: true})
			if err != nil {
				siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "refresh_token", Err: err})
				return
			}
			params.RefreshToken = value

		} else {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "refresh_token"})
			return
		}
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RefreshAccessToken(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/logout", wrapper.Logout)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/auth/oauth/google/callback", wrapper.CallbackFromGoogle)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/oauth/google/link", wrapper.LinkWithGoogle)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/oauth/google/login", wrapper.LoginWithGoogle)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/auth/refresh", wrapper.RefreshAccessToken)
	})

	return r
}